+++ src/spooq2/transformer/__init__.py
@@ -2,6 +2,7 @@ from .newest_by_group import NewestByGroup
 from .mapper import Mapper
 from .exploder import Exploder
 from .threshold_cleaner import ThresholdCleaner
+from .enum_cleaner import EnumCleaner
 from .sieve import Sieve
 from .flattener import Flattener
 
@@ -10,6 +11,7 @@ __all__ = [
     "Mapper",
     "Exploder",
     "ThresholdCleaner",
+    "EnumCleaner",
     "Sieve",
     "Flattener",
 ]