+++ src/spooq2/transformer/mapper_custom_data_types.py
@@ -829,7 +829,7 @@ def _generate_select_expression_for_extended_string_to_timestamp(source_column, 
             F.trim(source_column).cast(T.LongType()).cast(T.TimestampType())
         ).when(
             F.abs(F.trim(source_column).cast(T.LongType())) > MAX_TIMESTAMP_SEC,
-            (F.trim(source_column) / 1000).cast(T.LongType()).cast(T.TimestampType())
+            (F.trim(source_column) / 1000).cast(T.TimestampType())
         ).otherwise(F.trim(source_column).cast(T.TimestampType())).alias(name)
     )